Rationalise perl's OS-level signal handler
First some background:

UNIXy OSes support two type of signal handler function:

Signal_t handler1(int sig);
Signal_t handler3(int sig, siginfo_t *info, void *uap);

The original one-argument handler was set using the signal(2) system
call. The newer sigaction(2) system call allows either a 1-arg or
3-arg handler to be specified:

    act.sa_handler = handler1;
    sigaction(sig, act, NULL);

    act.sa_sigaction = handler3;
    act.sa_sa_flags |= SA_SIGINFO;
    sigaction(sig, act, NULL);

The current behaviour in perl core is that, in the presence of
HAS_SIGACTION and SA_SIGINFO, the signal handler type and function are
both declared as 3-arg, but perl still tells the kernel that the
supplied signal handler function takes one arg. This means that whenever
the kernel calls the handler, args 2 and 3 are whatever garbage the OS
and architecture cause them to happen to be. Furthermore, some code in
the handler uses the nullness of args 2 and 3 to decide whether it has
been called by the kernel, or explicitly as handler(sig,NULL,NULL).
The fact that this condition works has just been down to luck so far.

Note that POSIX.xs does allow a 3-arg signal handler to be specified by
passing the SA_SIGINFO flag, and a couple of tests check for this.

Recently, gcc-8 has been (quite reasonably) warning that we're passing
around 3-arg function pointers where a 1-arg function pointer is
expected.

This commit does the following:

1) It adds a new function, Perl_perly_sighandler() which is just
responsible for calling the perl-level $SIG{FOO} handler sub. This new
function may either be called directly by Perl_csighandler (for unsafe
signals), or called later by Perl_despatch_signals() for the more
normal safe signal regime. It has the 3 sig handler args, plus a fourth
boolean arg which indicates whether it is being called in safe mode.

2) It adds explicit 1-arg and 3-arg variants of the C-level signal
handler functions sighandler1() and sigghandler3(), while the existing
function sighandler() has a signature which depends on the value of
PERL_USE_3ARG_SIGHANDLER.

3) It disables PERL_USE_3ARG_SIGHANDLER by default. This means that
declarations such as Sighandler_t are now 1-arg to match the fact that
perl core is (as before) actually using a 1-arg handler most of the
time.

This commit isn't a complete fix: in particular using 3-arg handlers by
default by enabling PERL_USE_3ARG_SIGHANDLER doesn't yet pass all tests.
That is left as an exercise to the reader.

In summary, perl continues to use 1-arg signal handlers, but now
declares its handlers as 1-arg rather than 3-arg, so a bunch of compiler
warnings go away. Also, the code is now more rational and better
commented, so that moving to 3-arg handlers in the future won't be so
hard.